Revert "Merge remote-tracking branch 'origin/frameshift' into frameshift"
This reverts commit 3a2fb78712ffac73e28f5a36ad19d80efb2914e3, reversing
changes made to 32eb50ffcb3e39b8b190e80942e20d0d0c0b67ab.
