+++ Snakefile
@@ -141,10 +141,11 @@ rule duplicates_marking:
         " M={log.metrics}"
         " ASSUME_SORT_ORDER=coordinate 2> {log.stderr}"
 
-rule clusterrmdup:
+rule clusterrmdup_and_index:
     # Removes cluster duplicates and indexes output
     output:
         bam = "{dir}/mapped.sorted.tag.mkdup.bcmerge.bam",
+        bai = "{dir}/mapped.sorted.tag.mkdup.bcmerge.bam.bai",
         merges = "{dir}/barcode-merges.csv"
     input:
         bam = "{dir}/mapped.sorted.tag.mkdup.bam"
@@ -152,40 +153,30 @@ rule clusterrmdup:
     shell:
         "blr clusterrmdup"
         " {input.bam}"
-        " {output.bam}"
+        " -"
         " {output.merges}"
-        " -bc {config[cluster_tag]} 2>> {log}"
-
-rule buildmolecules:
-    # Groups reads into molecules depending on their genomic position and barcode
-    output:
-        bam = "{dir}/mapped.sorted.tag.mkdup.bcmerge.mol.bam"
-    input:
-        bam = "{dir}/mapped.sorted.tag.mkdup.bcmerge.bam"
-    log: "{dir}/buildmolecules.log"
-    shell:
-        "blr buildmolecules"
-        " {input.bam}"
-        " {output.bam}"
-        " -m {config[molecule_tag]}"
-        " -n {config[num_mol_tag]}"
-        " -bc {config[cluster_tag]}"
-        " 2> {log}"
+        " -bc {config[cluster_tag]} 2>> {log} |"
+        "tee {output.bam} |"
+        "samtools index - {output.bai}"
 
 rule filterclusters:
     # Filter clusters based on parameters
     output:
-        bam = "{dir}/mapped.sorted.tag.mkdup.bcmerge.mol.filt.bam",
+        bam = "{dir}/mapped.sorted.tag.mkdup.bcmerge.filt.bam",
+        stat1 = "{dir}/cluster_stats/bcmerge.stats.molecules_per_bc",
+        stat2 = "{dir}/cluster_stats/bcmerge.stats.molecule_stats"
     input:
-        bam = "{dir}/mapped.sorted.tag.mkdup.bcmerge.mol.bam"
+        bam = "{dir}/mapped.sorted.tag.mkdup.bcmerge.bam"
     log: "{dir}/filterclusters.log"
+    params:
+        stats = "{dir}/cluster_stats/bcmerge.stats"
     shell:
         "blr filterclusters"
-        " {input.bam}"
-        " {output.bam}"
-        " -mn {config[num_mol_tag]}"
         " -M 260"
-        " -t {config[cluster_tag]} {config[molecule_tag]} {config[num_mol_tag]} {config[sequence_tag]} 2> {log}"
+        " -s {params.stats}"
+        " -bc {config[cluster_tag]}"
+        " {input.bam}"
+        " {output.bam} 2>> {log}"
 
 rule bam_to_fastq:
     # Convert final bam file to fastq files for read 1 and 2
@@ -193,7 +184,7 @@ rule bam_to_fastq:
         r1_fastq = "{dir}/reads.1.final.fastq",
         r2_fastq = "{dir}/reads.2.final.fastq"
     input:
-        bam = "{dir}/mapped.sorted.tag.mkdup.bcmerge.mol.filt.bam"
+        bam = "{dir}/mapped.sorted.tag.mkdup.bcmerge.filt.bam"
     log: "{dir}/picard_samtofastq.log"
     params:
         picard_command = config["picard_command"],
