+++ tests/run.sh
@@ -11,5 +11,5 @@ ln -s $PWD/testdata/reads.2.fastq.gz outdir/reads.2.fastq.gz
 snakemake --configfile tests/test_config.yaml outdir/reads.1.final.fastq.gz \
     outdir/reads.2.final.fastq.gz
 
-m=$(samtools sort -n outdir/mapped.sorted.tag.mkdup.bcmerge.mol.filt.bam | samtools view - | md5sum | cut -f1 -d" ")
-test $m == 02595541aa2247266537cd772d712820
+m=$(samtools sort -n outdir/mapped.sorted.tag.mkdup.bcmerge.filt.bam | samtools view - | md5sum | cut -f1 -d" ")
+test $m == 0939535d1690f8733db59f3507490989
