+++ tests/test_config.yaml
@@ -1,8 +1,6 @@
 index_nucleotides: 1
 heap_space: 3
 cluster_tag: BC    # Used to store barcode cluster id in bam file. 'BX' is 10x genomic default
-molecule_tag: MI  # Used to store molecule ID, same as 10x default.
-num_mol_tag: MN # Used to store number of molecules per barcode
 sequence_tag: RX    # Used to store original barcode sequence in bam file. 'RX' is 10x genomic defaultbowtie2_reference: /Users/pontus.hojer/projects/BLR/testdata/chr1mini
 bowtie2_reference: ./testdata/chr1mini
 picard_command: picard
